Deserialize tilesets
